improved test
